requirements.txt
fastapi==0.104.0
uvicorn[standard]==0.24.0
gunicorn==21.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
azure-identity==1.12.0
azure-storage-blob==12.16.0
azure-ai-formrecognizer==3.3.0
pandas==2.0.0
streamlit==1.28.0
plotly==5.17.0
reportlab==4.0.0